Disabling verbose logging
Although it is helpful on some cituation to troubleshoot swift/memcache
related issues, this should only be turned on when needed. If one wants
to have it on by deafult, logratation needs to be setup for /var/log/memcache.log
otherwise the logging data could grow largely and fillup the partition.