Align partitions on optimized boundaries
Since more and more drives are coming with 4K sectors, a 0
starting position results in unaligned drives.

We could start the partition at a 4K boundary, but that still
makes parted whine because of alignment rules having to do with
flash devices.  If we start at 1M, all drives types are properly
aligned, parted doesn't complain, and we lose a reasonably
insignificant amount of disk space.

So that's what we'll do.

closes #136